Remove XlfFunctions - functions moved to XliffDocument
