+++ extension/src/LanguageFunctions.ts
@@ -8,16 +8,14 @@ import * as VSCodeFunctions from "./VSCodeFunctions";
 import * as escapeStringRegexp from "escape-string-regexp";
 import { XliffIdToken } from "./ALObject/XliffIdToken";
 import { Settings, Setting } from "./Settings";
-import {
-  targetStateActionNeededToken,
-  targetStateActionNeededKeywordList,
-} from "./Xliff/XlfFunctions";
 import * as Logging from "./Logging";
 import {
   CustomNoteType,
   StateQualifier,
   Target,
   TargetState,
+  targetStateActionNeededKeywordList,
+  targetStateActionNeededToken,
   TranslationToken,
   TransUnit,
   Xliff,
