+++ extension/src/Xliff/XLIFFDocument.ts
@@ -3,6 +3,8 @@
  */
 import * as fs from "fs";
 import * as xmldom from "xmldom";
+import * as escapeStringRegexp from "escape-string-regexp";
+
 
 import {
   XliffDocumentInterface,
@@ -17,7 +19,6 @@ import {
 } from "../XmlFormatter";
 import { isNullOrUndefined } from "util";
 import * as Common from "../Common";
-import { targetStateActionNeededAsList } from "./XlfFunctions";
 import * as LanguageFunctions from "../LanguageFunctions";
 import { XliffIdToken } from "../ALObject/XliffIdToken";
 
@@ -949,6 +950,7 @@ function compareTransUnitId(aUnit: TransUnit, bUnit: TransUnit): number {
   }
   return 0;
 }
+
 function transUnitIdAsObject(
   transUnit: TransUnit
 ): { objectTypeId: number; controlId: number; propertyId: number } {
@@ -964,4 +966,52 @@ function transUnitIdAsObject(
     controlId: parseInt(fieldId),
     propertyId: parseInt(propertyId),
   };
+}
+
+function targetStateActionNeededAsList(
+  lowerThanTargetState?: TargetState
+): string[] {
+  const stateActionNeeded = [
+    TargetState.needsAdaptation,
+    TargetState.needsL10n,
+    TargetState.needsReviewAdaptation,
+    TargetState.needsReviewL10n,
+    TargetState.needsReviewTranslation,
+    TargetState.needsTranslation,
+    TargetState.new,
+  ];
+  if (lowerThanTargetState) {
+    switch (lowerThanTargetState) {
+      case TargetState.signedOff:
+        stateActionNeeded.push(TargetState.translated);
+        break;
+      case TargetState.final:
+        stateActionNeeded.push(TargetState.translated);
+        stateActionNeeded.push(TargetState.signedOff);
+        break;
+    }
+  }
+  return stateActionNeeded;
+}
+
+export function targetStateActionNeededKeywordList(
+  lowerThanTargetState?: TargetState
+): Array<string> {
+  const keywordList: Array<string> = [];
+  targetStateActionNeededAsList(lowerThanTargetState).forEach((s) => {
+    keywordList.push(`state="${s}"`);
+  });
+  return keywordList;
+}
+
+export function targetStateActionNeededToken(): string {
+  return (
+    `state="${escapeStringRegexp(TargetState.needsAdaptation)}"|` +
+    `state="${escapeStringRegexp(TargetState.needsL10n)}"|` +
+    `state="${escapeStringRegexp(TargetState.needsReviewAdaptation)}"|` +
+    `state="${escapeStringRegexp(TargetState.needsReviewL10n)}"|` +
+    `state="${escapeStringRegexp(TargetState.needsReviewTranslation)}"|` +
+    `state="${escapeStringRegexp(TargetState.needsTranslation)}"|` +
+    `state="${escapeStringRegexp(TargetState.new)}"`
+  );
 }
